requirements.txt
flask>=2.2
python-dotenv>=1.0
requests>=2.28
transformers>=4.27
torch
flask-sqlalchemy>=3.0